requirements.txt
# Core dependencies
openai>=1.0.0
requests>=2.25.0
pydantic>=2.0.0
python-dotenv>=1.0.0
langchain>=0.1.0

# Web scraping and HTTP requests
beautifulsoup4>=4.12.0
httpx>=0.24.0
lxml>=4.9.0

# Video processing and multimedia
imageio-ffmpeg>=0.4.8
imageio>=2.31.0

# Speech recognition and subtitle generation
faster-whisper>=0.10.0
srt>=3.5.0

# Additional utilities
tqdm>=4.65.0

# Standard library extensions (usually included with Python)
# typing - built-in
# os - built-in
# time - built-in
# shutil - built-in
# random - built-in
# urllib.parse - built-in
# json - built-in
# datetime - built-in
# re - built-in
